Move load_engine Out of a Closed Benchmark and Into costbench
`load_engine` lived in `scripts/bench_bitplanes.py`, a targeted benchmark whose
own investigation closed (all four of its doc references are under
`docs/issues/done/`), and 35 call sites across the toolkit imported it from
there. Nothing about it is bitplane-specific.

Moves it and `BATCH_SIZE` to `costbench`, repoints every importer, and drops
`card_engine` to a type-checking import in `bench_bitplanes` now that only an
annotation needs it. `bench_bitplanes` keeps `bench_one` / `WARMUP`, which are
genuinely its own.

Two of the scripts on the audit's delete list -- `bench_collection_eq_gt` and
`bench_permuted_order` -- still carry private copies of `load_engine`. Left
alone; they go away with the rest.

All 42 bench/study/census scripts import cleanly afterwards. The exception is
`bench_random`, which has no `__main__` guard and runs its whole benchmark on
import; that is pre-existing and untouched here, and it is on the delete list.
