+++ scripts/bench_cost_error_attribution.py
@@ -45,7 +45,7 @@ REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(REPO_ROOT))
 
 import scripts.fit_cost_model as fitmod  # noqa: E402
-from scripts.bench_bitplanes import load_engine  # noqa: E402
+from scripts.costbench import load_engine  # noqa: E402
 from scripts.fit_cost_model import collect, fit_log_ratio  # noqa: E402
 from scripts.query_sampler import MODES, QuerySampler  # noqa: E402
