+++ scripts/bench_cost_model_agreement.py
@@ -53,7 +53,7 @@ sys.path.insert(0, str(REPO_ROOT))
 
 from api.parsing import parse_scryfall_query  # noqa: E402
 from scripts import costbench  # noqa: E402
-from scripts.bench_bitplanes import load_engine  # noqa: E402
+from scripts.costbench import load_engine  # noqa: E402
 
 NUM_WARMUPS = 2
 NUM_TRIALS = 7
