+++ scripts/bench_name_order.py
@@ -24,7 +24,8 @@ import sys
 REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(REPO_ROOT))
 
-from scripts.bench_bitplanes import bench_one, load_engine  # noqa: E402
+from scripts.bench_bitplanes import bench_one  # noqa: E402
+from scripts.costbench import load_engine  # noqa: E402
 
 # Each config row is (group, query, unique, orderby, prefer).
 CONFIGS: list[tuple[str, str, str, str, str]] = [
