+++ scripts/bench_oracle_word_index.py
@@ -28,7 +28,8 @@ REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(REPO_ROOT))
 
 from api.parsing import parse_scryfall_query  # noqa: E402
-from scripts.bench_bitplanes import bench_one, load_engine  # noqa: E402
+from scripts.bench_bitplanes import bench_one  # noqa: E402
+from scripts.costbench import load_engine  # noqa: E402
 
 # (group, query, unique, orderby, prefer) — direction=asc, limit=100, offset=0 throughout.
 CONFIGS: list[tuple[str, str, str, str, str]] = [
