+++ scripts/bench_text_memo.py
@@ -21,7 +21,8 @@ import sys
 REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(REPO_ROOT))
 
-from scripts.bench_bitplanes import bench_one, load_engine  # noqa: E402
+from scripts.bench_bitplanes import bench_one  # noqa: E402
+from scripts.costbench import load_engine  # noqa: E402
 
 # (group, query) — unique=card, orderby=edhrec, prefer=default throughout.
 CONFIGS: list[tuple[str, str]] = [
