+++ scripts/fit_cost_model.py
@@ -37,8 +37,8 @@ sys.path.insert(0, str(REPO_ROOT))
 
 from api.parsing import parse_scryfall_query  # noqa: E402
 from scripts import costbench  # noqa: E402
-from scripts.bench_bitplanes import load_engine  # noqa: E402
 from scripts.bench_cost_model_agreement import AGREE_HI, AGREE_LO  # noqa: E402
+from scripts.costbench import load_engine  # noqa: E402
 from scripts.query_sampler import QuerySampler  # noqa: E402
 
 NUM_WARMUPS = 2
